ask-maas-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
redis[async]==5.0.7
faiss-cpu==1.8.0
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.0
beautifulsoup4==4.12.3
lxml==5.2.2
markdown==3.6
python-multipart==0.0.9
sse-starlette==2.1.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-httpx==0.46b0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-prometheus==0.46b0
prometheus-client==0.20.0
aiofiles==23.2.1
PyGithub==2.3.0
tiktoken==0.7.0
rank-bm25==0.2.2
tenacity==8.4.2
python-jose[cryptography]==3.3.0
structlog==24.2.0
python-dotenv==1.0.1
asyncio-throttle==1.0.2
